Updates: - Add pyproject.toml to build the package - Add examples   * Freesurfer recon-all small example   * [WIP] LivingPark replication paper (hanganu-etal-2014)
