+++ .gitmodules
@@ -1,0 +1,3 @@
+[submodule "example/hanganu-etal-2014"]
+	path = example/hanganu-etal-2014
+	url = https://github.com/yohanchatelain/hanganu-etal-2014
