--- __init__.py
@@ -1,10 +1,0 @@
-"""
-VIP PACKAGE
-
-Implement the API 'Carmin' that is used to interact with VIP.
-Contact Sorina Camarasu-Pop and Axel Bonnet for more informations.
-"""
-
-# info
-__version__ = "1.0"
-__author__ = "Timothée Chabat"
