+++ example/freesurfer/README.md
@@ -1,0 +1,15 @@
+# Small example using VIP API
+
+`./freesurfer-recon-all.sh` script launches Freesurfer on VIP server through VIP API.
+
+## Requirements
+
+### VIP API token
+
+Export the API token before executing `./freesurfer-recon-all.sh`
+
+`export VIP_API_TOKEN=<token>`
+
+### Freesurfer license
+
+Add your Freesurfer license in `inputs/LICENSE.txt`
