+++ example/freesurfer/freesurfer-args.json
@@ -1,0 +1,6 @@
+{
+    "image": "/vip/Home/API/reconall-example/INPUTS/sub-1_T1w.nii.gz",
+    "results-directory": "/vip/Home/API/reconall-example/OUTPUTS/",
+    "options": "-all",
+    "license key": "/vip/Home/API/reconall-example/INPUTS/LICENSE.txt"
+}
